#RealEstate.ts #S3 correction #CompteHeader props #AjoutBienScreen textInput recovery values except TaxType, button functionality

+++ utils/S3FileStorage.ts
@@ -12,9 +12,8 @@ import { useEffect } from 'react';
 import { ImagePickerResult } from 'expo-image-picker';
 import { DocumentResult } from 'expo-document-picker';
 import { v4 as uuid } from 'uuid';
+import * as mime from 'react-native-mime-types';
 import { useNetworkInfo } from './CustomHooks';
-
-const MimeType = require('mime-type/with-db');
 
 export const waitingDirectory = 'waitingFile/';
 
@@ -48,8 +47,14 @@ const UploadInternal = async (
   try {
     const response = await fetch(uri);
     const blob = await response.blob();
-    const fileInfo = await FileSystem.getInfoAsync(uri);
-    if (fileInfo.exists) {
+    let exists = false;
+    if (Platform.OS === 'web') {
+      exists = true;
+    } else {
+      const fileInfo = await FileSystem.getInfoAsync(uri);
+      exists = fileInfo.exists;
+    }
+    if (exists) {
       await Storage.put(key, blob, {
         contentType: blob.type,
         progressCallback(progress: any) {
@@ -57,7 +62,7 @@ const UploadInternal = async (
         },
       });
     }
-    if (deleteTempFile) {
+    if (deleteTempFile && Platform.OS !== 'web') {
       await FileSystem.deleteAsync(uri);
     }
     if (storageKey) {
@@ -103,7 +108,7 @@ export const Upload = async (file: ImagePickerResult | DocumentResult, path?: st
     return false;
   }
 
-  const extension = finalFile.name.split('.');
+  const extension = finalFile.name.split('.').pop();
   const uuidKey = uuid();
   const key = `${path}${uuidKey}.${extension}`;
   const inputData = {
@@ -135,7 +140,7 @@ export const Upload = async (file: ImagePickerResult | DocumentResult, path?: st
 
 const getFilename = (filePath: string) => {
   if (filePath.indexOf('data:') > -1) {
-    const ext = MimeType.extension(filePath.split(';')[0].replace('data:', ''));
+    const ext = mime.extension(filePath.split(';')[0].replace('data:', ''));
     return `${Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15)}.${ext}`;
   }
   return filePath.replace(/^.*[\\/]/, '');
